+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * [BUGFIX] store event objects as hashes so they can be safely serialized using JSON (default since rails 4.1) #13
   * added support for Criteo #17
   * added support for Google Analytics User ID feature #15
+  * added support for Google Tag Manager feature #29
 
 # 0.3.0